% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgf.R
\name{criticalValue}
\alias{criticalValue}
\title{Critical value x such that \eqn{P(LRT > x) \le p}}
\usage{
criticalValue(k, n, p = 0.05, verbose = FALSE)
}
\arguments{
\item{k}{number of categories (a vector for independent multinomial draws)}

\item{n}{sample size (a vector for independent multinomial draws)}

\item{p}{significance level (e.g., 0.05)}

\item{verbose}{draw the minimizer if \code{TRUE}}
}
\description{
Critical value x such that \eqn{P(LRT > x) \le p}
}
\examples{
n <- 1:100
crit <- sapply(n, function(.n) criticalValue(20, .n, p=0.01))
plot(n, crit)
abline(h=qchisq(0.01, df=20-1, lower.tail = FALSE)) # standard chi-square asymptotic
criticalValue(20, 100, p=0.01, verbose=TRUE)
}
\seealso{
tailProbBound
}
