% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgf.R
\name{mgfBound}
\alias{mgfBound}
\title{G(lambda; k,n) upper bound of the moment generating function on (LRT / 2)}
\usage{
mgfBound(k, n, lambda)
}
\arguments{
\item{k}{number of categories}

\item{n}{sample size}

\item{lambda}{number between 0 and 1}
}
\description{
G(lambda; k,n) upper bound of the moment generating function on (LRT / 2)
}
