% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgf.R
\name{tailProbBound}
\alias{tailProbBound}
\title{Tail bound on P(LRT > x)}
\usage{
tailProbBound(x, k, n, verbose = FALSE)
}
\arguments{
\item{x}{the value of the LRT}

\item{k}{number of categories (a vector for independent multinomial draws)}

\item{n}{sample size (a vector for independent multinomial draws)}

\item{verbose}{draw the minimizer if \code{TRUE}}
}
\value{
an upper bound on P(LRT > x), which can be used as a conservative p-value
}
\description{
Tail bound on P(LRT > x)
}
\note{
For independent multinomial samples, k and n must be of the same length.
}
\examples{
tailProbBound(20, 7, 50)
pchisq(20, 6, lower.tail=FALSE) # compare with the standard chi-square asymptotic
tailProbBound(30, c(5,10), c(20,50))
pchisq(20, 13, lower.tail=FALSE) # compare with the standard chi-square asymptotic
}
